% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econometrics.R
\name{choose_model}
\alias{choose_model}
\title{Choose model}
\usage{
choose_model()
}
\value{
model_choice
}
\description{
Allows the user to choose one or all of several econometric models to be
calculated by run_models.
}
\author{
Michael David Gill
}

