% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econometrics.R
\name{construct_configuration_df}
\alias{construct_configuration_df}
\title{Construct Configuration Data Frame}
\usage{
construct_configuration_df()
}
\value{
configuration_df a data frame of MySQL configuration parameters.
}
\description{
A helper function that constructs a data frame of configuration data for a
MySQL server from a file or user input for use by the load_sql_database
function.
}
\author{
Michael David Gill
}

