% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econometrics-package.R
\docType{package}
\name{econometrics-package}
\alias{econometrics}
\alias{econometrics-package}
\title{econometrics: Datasets and Unofficial Solutions in C++ for the Exercises from 'Econometrics' by Bruce E. Hansen}
\description{
'Econometrics' by Bruce E. Hansen <isbn:9780691235899> includes exercises that are solved in R. This package provides unofficial solutions in C++ for the exercises in the book.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pacha.dev/cpp11armadillo}
  \item Report bugs at \url{https://github.com/pachadotdev/cpp11armadillo/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

}
\keyword{internal}
