% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econometrics-package.R
\name{influence_dbl}
\alias{influence_dbl}
\title{Influence}
\usage{
influence_dbl(y, X)
}
\arguments{
\item{y}{numeric vector of dependent variable}

\item{X}{numeric matrix of independent variables}
}
\description{
Influence
}
