% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econometrics.R
\name{load_sql_database}
\alias{load_sql_database}
\title{Load SQL Database}
\usage{
load_sql_database()
}
\value{
mysql_db a SQL database connection.
}
\description{
Loads a MySQL database.
}
\author{
Michael David Gill
}

