% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econometrics-package.R
\name{ols_mat}
\alias{ols_mat}
\title{OLS Estimator}
\usage{
ols_mat(y, X)
}
\arguments{
\item{y}{numeric vector of dependent variable}

\item{X}{numeric matrix of independent variables}
}
\description{
OLS Estimator
}
