% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econometrics.R
\name{r_sql_econometrics}
\alias{r_sql_econometrics}
\title{R SQL Econometrics}
\usage{
r_sql_econometrics()
}
\description{
Runs various pooled linear models and tests for econometric analysis using
data imported from a MySQL server.
}
\details{
Sets all parameters at the start of the script for the further steps (file
path, predefine which regressions, tests and trimming should be performed, if
SQL Script is performed or not, level of trimming, set y, x’s, fixed effects,
instrument variables, …)

Connects to MySQL database (e.g. ‘RMySQL’)
- Credentials are available from a text file or likewise
- User: root / password: test / host: localhost / port:3306

Runs SQL Script from file and wait until finished

Imports all data from database to R (schema: test / table: data)

Waits 10 min and then shutdown MySQL database (e.g. SQL command: SHUTDOWN;)
}
\author{
Michael David Gill
}

