% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econometrics.R
\name{run_models}
\alias{run_models}
\title{Run Models}
\usage{
run_models(y, x, x_fe, x_iv, data, model_choice)
}
\arguments{
\item{y}{column numbers of outcome variable to be predicted.}

\item{x}{a vector of column numbers of predictor/covariate variable(s).}

\item{x_fe}{optional vector of column numbers of fixed effects.}

\item{x_iv}{optional vector of column numbers of instrumental variables.}

\item{data}{data frame to be modeled.}

\item{model_choice}{type of model to be run, either (1) pooled OLS, (2) fixed effects, (3) random
effects, (4) regression with instrumental variables, or (5) regression with
fixed effects and instrumental variables.}
}
\value{
model
}
\description{
Calculates one or several econometric models.
}
\details{
description
Runs a pooled OLS and save the estimates with ‘plm’ function (-->pooling), a
random regression and save the estimates with ‘plm’ function (-->random), a
fixed effects regression  and save the estimates with ‘plm’ function
(-->fixed), a regression with instrumental variables (-->IV) ivreg {AER}, and
a regression with instrumental variables and fixed effects (--> fixed_IV)
}
\author{
Michael David Gill
}
\references{
Croissant, Y., & Millo, G. (2008). Panel data econometrics in R:
The plm package. \emph{Journal of Statistical Software, 27}(2), 1–43.
http://doi.org/10.18637/jss.v027.i02
}

