% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econometrics.R
\name{run_sql_script}
\alias{run_sql_script}
\title{Run SQL Script}
\usage{
run_sql_script(mysql_db)
}
\arguments{
\item{mysql_db}{name of database.}
}
\value{
data_frame
resulting table from SQL script formatted as an R data frame.
}
\description{
Runs a SQL query imported from an SQL query file.
}
\author{
Michael David Gill
}

