% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econometrics.R
\name{run_tests}
\alias{run_tests}
\title{Run Tests}
\usage{
run_tests(model_list)
}
\arguments{
\item{model_list}{a list of panel data estimator models.}
}
\value{
a list of test results.
}
\description{
Runs several test on models generated by run_models.
}
\details{
Tests for multicollinearity by calculating variance inflation factors (VIF)
for each variable, pooled OLS vs. random effects with Breusch-Pagan Lagrange
multiplier (plmtest), time-fixed effects: (plmtest), heteroskedasticity
(bptest / White’s test), Random vs. fixed effects: Hausman test (phtest),
time-fixed effects (pFtest), weak instrument (F-test / Wald test).
}
\author{
Michael David Gill
}

