% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econometrics.R
\name{summarize_data}
\alias{summarize_data}
\title{Summarize Data}
\usage{
summarize_data(data_frame, left_trim, right_trim)
}
\arguments{
\item{data_frame}{data frame to be summarized.}

\item{left_trim}{percentage of observations to trim from right of the variable's distribution.}

\item{right_trim}{percentage of observations to trim from left of the variable's distribution.}
}
\value{
summary
data frame including mean, median, standard deviation, minimum, and maximum
for each variable.
}
\description{
Calculates summary of descriptive statisitics (mean, standard deviation, min,
max, correlation).
}
\author{
Michael David Gill
}

