\name{econometrics-package}
\alias{econometrics-package}
\alias{econometrics}
\docType{package}
\title{
\packageTitle{econometrics}
}
\description{
\packageDescription{econometrics}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{econometrics}
\packageIndices{econometrics}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{econometrics}

Maintainer: \packageMaintainer{econometrics}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
